requirements.txt
pandas
numpy
matplotlib
kafka-python